rename test_aux_calcs.py
